Move extension-related classes to separate package (issue #242).
