+++ plugins/org.franca.deploymodel.dsl.ui/src/org/franca/deploymodel/dsl/ui/contentassist/FDeployProposalProvider.java
@@ -37,17 +37,16 @@ import org.franca.core.franca.FTypeCollection;
 import org.franca.core.franca.FUnionType;
 import org.franca.core.utils.FrancaIDLUtils;
 import org.franca.deploymodel.core.FDModelUtils;
-import org.franca.deploymodel.dsl.ExtensionRegistry;
-import org.franca.deploymodel.dsl.IFDeployExtension;
-import org.franca.deploymodel.dsl.IFDeployExtension.Root;
 import org.franca.deploymodel.dsl.fDeploy.FDBuiltInPropertyHost;
-import org.franca.deploymodel.dsl.fDeploy.FDDeclaration;
 import org.franca.deploymodel.dsl.fDeploy.FDModel;
 import org.franca.deploymodel.dsl.fDeploy.FDOverwriteElement;
 import org.franca.deploymodel.dsl.fDeploy.FDeployPackage;
 import org.franca.deploymodel.dsl.fDeploy.Import;
 import org.franca.deploymodel.dsl.scoping.DeploySpecProvider;
 import org.franca.deploymodel.dsl.scoping.DeploySpecProvider.DeploySpecEntry;
+import org.franca.deploymodel.extensions.ExtensionRegistry;
+import org.franca.deploymodel.extensions.IFDeployExtension;
+import org.franca.deploymodel.extensions.IFDeployExtension.Root;
 
 import com.google.common.base.Joiner;
 import com.google.common.collect.Lists;
