+++ plugins/org.franca.deploymodel.dsl/src/org/franca/deploymodel/extensions/ExtensionRegistry.xtend
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.franca.deploymodel.dsl
+package org.franca.deploymodel.extensions
 
 import com.google.common.collect.Lists
 import java.util.Collection
