+++ plugins/org.franca.deploymodel.dsl/src/org/franca/deploymodel/extensions/IFDeployExtension.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  *******************************************************************************/
-package org.franca.deploymodel.dsl;
+package org.franca.deploymodel.extensions;
 
 import java.util.Collection;
