+++ plugins/org.franca.deploymodel.ext.providers/src/org/franca/deploymodel/ext/providers/ProviderExtension.xtend
@@ -1,6 +1,6 @@
 package org.franca.deploymodel.ext.providers
 
-import org.franca.deploymodel.dsl.IFDeployExtension
+import org.franca.deploymodel.extensions.IFDeployExtension
 import java.util.Collection
 
 class ProviderExtension implements IFDeployExtension {
